changes part 1

+++ src/main/java/com/mmt/user/services/UserService.java
@@ -4,9 +4,6 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import com.mmt.address.dao.AddressDao;
-import com.mmt.address.Address;
 import com.mmt.bookedFlight.dao.BookedFlightDao;
 import com.mmt.bookedFlight.model.BookedFlight;
 import com.mmt.bookedHotel.dao.BookedHotelDao;
@@ -27,8 +24,6 @@ public class UserService implements UserServiceInterface{
 	@Autowired
 	private BookedHotelDao bhd;
 	
-	@Autowired
-	private AddressDao ad;
 	
 	@Override
 	public List<User> allUsers() {
@@ -43,7 +38,9 @@ public class UserService implements UserServiceInterface{
 
 	@Override
 	public boolean userLogin(User user) {
-		return ud.findByEmailIdAndPassword(user.getEmailId(), user.getPassword());
+		User existUser= ud.findByEmailIdAndPassword(user.getEmailId(), user.getPassword());
+		if(existUser==null) return false;
+		return true;
 	}
 
 	@Override
@@ -53,21 +50,23 @@ public class UserService implements UserServiceInterface{
 	}
 
 	@Override
-	public boolean deleteUser(String userId) {
-		ud.deleteById(userId);
-		return false;
+	public boolean deleteUser(String userId, String password) {
+		User existUser= ud.findByUserIdAndPassword(userId, password);
+		if(existUser==null) return false;
+			ud.deleteById(userId);
+			return true;
 	}
 
 	@Override
-	public boolean updateUser(User user) {
+	public boolean updateUser(User user, String userId) {
 	
 		User existUser = ud.findById(user.getUserId()).get();
 		if(user.getMobileNumber() != null) {
 			existUser.setMobileNumber(user.getMobileNumber());
 		}
-		if(user.getAddress()!=null) {
-			existUser.setAddress(user.getAddress());
-		}
+		/*
+		 * if(user.getAddress()!=null) { existUser.setAddress(user.getAddress()); }
+		 */
 		if(user.getFirstName()!=null) {
 			existUser.setFirstName(user.getFirstName());
 		}
@@ -133,21 +132,26 @@ public class UserService implements UserServiceInterface{
 		return bfd.findByUser(user);
 	}
 
+	/*
+	 * @Override public boolean updateUserAddress(Address address, String userId) {
+	 * User user = ud.findById(userId).get(); Address oldAddress =
+	 * user.getAddress(); if(address.getArea()!=null)
+	 * oldAddress.setArea(address.getArea()); if(address.getCity()!=null)
+	 * oldAddress.setCity(address.getCity());
+	 * if(address.getCountry()!=null)oldAddress.setCountry(address.getCountry());
+	 * if(address.getHouseNo()!=null) oldAddress.setHouseNo(address.getHouseNo());
+	 * if(address.getPinCode()==0) oldAddress.setPinCode(address.getPinCode());
+	 * if(address.getState()!=null) oldAddress.setState(address.getState());
+	 * if(address.getStreetName()!=null)
+	 * oldAddress.setStreetName(address.getStreetName());
+	 * user.setAddress(oldAddress); ud.save(user); ad.save(oldAddress); return
+	 * false; }
+	 */
+
 	@Override
-	public boolean updateUserAddress(Address address, String userId) {
-		User user = ud.findById(userId).get();
-		Address oldAddress = user.getAddress();
-		if(address.getArea()!=null) oldAddress.setArea(address.getArea());
-		if(address.getCity()!=null) oldAddress.setCity(address.getCity());
-		if(address.getCountry()!=null)oldAddress.setCountry(address.getCountry());
-		if(address.getHouseNo()!=null) oldAddress.setHouseNo(address.getHouseNo());
-		if(address.getPinCode()==0) oldAddress.setPinCode(address.getPinCode());
-		if(address.getState()!=null) oldAddress.setState(address.getState());
-		if(address.getStreetName()!=null) oldAddress.setStreetName(address.getStreetName());
-		user.setAddress(oldAddress);
-		ud.save(user);
-		ad.save(oldAddress);
-		return false;
+	public String userName(String email, String password) {
+		User user = ud.findByEmailIdAndPassword(email, password);
+		return user.getUserId();
 	}
